moved a folder up to components directory
