+++ src/components/header/Header.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/mixins';
+@import '../../assets/scss/mixins';
 
 .header {
 	padding: 6rem 0;
