+++ src/components/project/Project.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/mixins';
+@import '../../assets/scss/mixins';
 .project {
 	padding: 3rem 0;
 	display: flex;
