+++ src/components/social/Social.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/mixins';
+@import '../../assets/scss/mixins';
 
 .social {
 	display: flex;
