+++ src/components/spinner/Spinner.scss
