+++ src/components/stack/Stack.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/mixins';
+@import '../../assets/scss/mixins';
 .stack {
 	padding: 6rem 0;
