+++ src/components/typed/TypedComponent.js
